google-cloud-related/cloud-storage-operations/requirements.txt
google-cloud-storage==2.7.0



